+++ src/vallenae/_pyinstaller/hook-vallenae.py
@@ -1,0 +1,5 @@
+# pylint: disable=invalid-name
+
+from PyInstaller.utils.hooks import collect_data_files
+
+datas = collect_data_files("vallenae", subdir="io/schema_templates", includes=["*.sql"])