+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Example for custom feature extraction
+- PyInstaller hook
 
 ## [0.6.0] - 2021-09-02