+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2022-11-10
+
 ### Added
 
 - Example for custom feature extraction
 - PyInstaller hook
+- CI for Python 3.10
 
 ### Changed
 
@@ -147,7 +150,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial public release
 
-[Unreleased]: https://github.com/vallen-systems/pyVallenAE/compare/0.6.0...HEAD
+[Unreleased]: https://github.com/vallen-systems/pyVallenAE/compare/0.7.0...HEAD
+[0.7.0]: https://github.com/vallen-systems/pyVallenAE/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/vallen-systems/pyVallenAE/compare/0.5.4...0.6.0
 [0.5.4]: https://github.com/vallen-systems/pyVallenAE/compare/0.5.3...0.5.4
 [0.5.3]: https://github.com/vallen-systems/pyVallenAE/compare/0.5.2...0.5.3